archive/legacy_scripts/requirements.txt
pandas
pylcs
openpyxl
pymorphy3
pymorphy3-dicts-ru
pyspellchecker
nltk